tests: Move fill blocks test to RPC location
