+++ testing/integration-tests/src/client/mod.rs
@@ -11,13 +11,17 @@ use crate::{
         wait_for_blocks,
     },
 };
+use futures::StreamExt;
 use sp_core::{
     sr25519::Pair as Sr25519Pair,
     storage::well_known_keys,
     Pair,
 };
 use sp_keyring::AccountKeyring;
-use subxt::error::DispatchError;
+use subxt::{
+    error::DispatchError,
+    rpc::Rpc,
+};
 
 #[tokio::test]
 async fn insert_key() {
@@ -89,6 +93,51 @@ async fn chain_subscribe_finalized_blocks() {
 
     let mut blocks = api.rpc().subscribe_finalized_blocks().await.unwrap();
     blocks.next().await.unwrap().unwrap();
+}
+
+#[tokio::test]
+async fn missing_block_headers_will_be_filled_in() -> Result<(), subxt::Error> {
+    let ctx = test_context().await;
+    let api = ctx.client();
+
+    // Manually subscribe to the next 6 finalized block headers, but deliberately
+    // filter out some in the middle so we get back b _ _ b _ b. This guarantees
+    // that there will be some gaps, even if there aren't any from the subscription.
+    let some_finalized_blocks = api
+        .rpc()
+        .subscribe_finalized_blocks()
+        .await?
+        .enumerate()
+        .take(6)
+        .filter(|(n, _)| {
+            let n = *n;
+            async move { n == 0 || n == 3 || n == 5 }
+        })
+        .map(|(_, h)| h);
+
+    // This should spot any gaps in the middle and fill them back in.
+    let all_finalized_blocks = Rpc::subscribe_to_block_headers_filling_in_gaps(
+        ctx.client().rpc().clone(),
+        None,
+        some_finalized_blocks,
+    );
+    futures::pin_mut!(all_finalized_blocks);
+
+    // Iterate the block headers, making sure we get them all in order.
+    let mut last_block_number = None;
+    while let Some(header) = all_finalized_blocks.next().await {
+        let header = header?;
+
+        use sp_runtime::traits::Header;
+        let block_number: u128 = (*header.number()).into();
+
+        if let Some(last) = last_block_number {
+            assert_eq!(last + 1, block_number);
+        }
+        last_block_number = Some(block_number);
+    }
+
+    Ok(())
 }
 
 #[tokio::test]
